refactor: 불필요한 Optional 제거
