+++ backend/src/main/java/com/techeerlog/member/controller/MemberController.java
@@ -12,37 +12,30 @@ import com.techeerlog.member.dto.request.SignupRequest;
 import com.techeerlog.member.dto.request.UpdatePasswordRequest;
 import com.techeerlog.member.dto.response.MemberResponse;
 import com.techeerlog.member.dto.response.ProfileResponse;
-import com.techeerlog.member.service.MemberService;
+import com.techeerlog.member.service.MemberServiceImpl;
 import io.swagger.v3.oas.annotations.Operation;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.util.Optional;
 
 import static com.techeerlog.global.response.ResultCode.*;
 
 
 @RestController
 @RequestMapping("/v1/members")
+@RequiredArgsConstructor
 public class MemberController {
 
-    private final MemberService memberService;
+    private final MemberServiceImpl memberServiceImpl;
     private final MemberMapper memberMapper;
     private final RefreshTokenService refreshTokenService;
-
-
-    public MemberController(MemberService memberService, RefreshTokenService refreshTokenService, MemberMapper memberMapper) {
-        this.memberService = memberService;
-        this.refreshTokenService = refreshTokenService;
-        this.memberMapper = memberMapper;
-    }
 
     @Operation(summary = "회원가입", description = "회원가입 기능")
     @PostMapping(value = "/signup")
     public ResponseEntity<ResultResponse<MemberResponse>> signUp(@RequestBody SignupRequest signupRequest) {
-        Member member = memberService.signUp(signupRequest);
-        ResultResponse<MemberResponse> resultResponse = new ResultResponse<>(SIGNUP_SUCCESS, memberMapper.memberToMemberResponse(member));
+        Member memberResponse = memberServiceImpl.signUp(signupRequest);
+        ResultResponse<MemberResponse> resultResponse = new ResultResponse<>(SIGNUP_SUCCESS, memberMapper.memberToMemberResponse(memberResponse));
 
         return ResponseEntity.status(SIGNUP_SUCCESS.getStatus())
                 .body(resultResponse);
@@ -51,7 +44,7 @@ public class MemberController {
     @Operation(summary = "프로필 조회", description = "프로필 조회 기능")
     @GetMapping("/profile")
     public ResponseEntity<ResultResponse<ProfileResponse>> findProfile(@Login AuthInfo authInfo) {
-        ProfileResponse profileResponse = memberService.findProfile(authInfo);
+        ProfileResponse profileResponse = memberServiceImpl.findProfile(authInfo);
         ResultResponse<ProfileResponse> resultResponse = new ResultResponse<>(FIND_PROFILE_SUCCESS, profileResponse);
 
         return ResponseEntity.status(FIND_PROFILE_SUCCESS.getStatus())
@@ -60,11 +53,11 @@ public class MemberController {
 
     @Operation(summary = "프로필 수정", description = "프로필 수정 기능")
     @PatchMapping(consumes = "multipart/form-data")
-    public ResponseEntity<SimpleResultResponse> editMember(@RequestPart(value = "data", required = false) EditMemberRequest editMemberRequest,
-                                                           @RequestPart(value = "part", required = false) Optional<MultipartFile> multipartFile,
+    public ResponseEntity<SimpleResultResponse> editMember(@RequestPart(value = "data") EditMemberRequest editMemberRequest,
+                                                           @RequestPart(value = "part", required = false) MultipartFile multipartFile,
                                                            @Login AuthInfo authInfo) {
 
-        memberService.editMemberInfo(editMemberRequest, authInfo, multipartFile);
+        memberServiceImpl.editMemberInfo(editMemberRequest, authInfo, multipartFile);
         refreshTokenService.deleteToken(authInfo.getId());
 
         SimpleResultResponse resultResponse = new SimpleResultResponse(EDIT_PROFILE_SUCCESS);
@@ -77,7 +70,7 @@ public class MemberController {
     @PatchMapping("/password")
     public ResponseEntity<SimpleResultResponse> updatePassword(@RequestBody UpdatePasswordRequest updatePasswordRequest,
                                                                @Login AuthInfo authInfo) {
-        memberService.updatePassword(authInfo, updatePasswordRequest);
+        memberServiceImpl.updatePassword(authInfo, updatePasswordRequest);
         refreshTokenService.deleteToken(authInfo.getId());
         SimpleResultResponse resultResponse = new SimpleResultResponse(UPDATE_CODE_SUCCESS);
